// src/enums/common.enums.ts
export enum PagingAction {
    First = 'First',
    Last = 'Last',
    Prev = 'Prev',
    Next = 'Next',
    Page = 'Page',
    All = 'All'
}

export enum RoleEnum {
    Admin = 'admin',
    CompanyHead = 'companyadmin',
    PlantHead = 'plantleader',
    LineHead = 'lineleader',
    DeviceHead = 'deviceleader',
}

export enum RightsEnum {
    Create='create',
    View='view',
    Edit='edit'
}

export enum ModuleNameEnum {
    Device = 'Device',
    Line = 'Line',
    Plant = 'Plant',
    Navigation = 'Navigation',
    User = 'User'
}

export enum MenuNameEnum {
    DeviceView = 'Device View',
    DeviceCreate = 'Device Create',
    LineView = 'Line View',
    LineCreate = 'Line Create',
    PlantView = 'Plant View',
    PlantCreate = 'Plant Create',
    Navigation = 'Navigation Role Wise',
    UserView = 'User View',
    UserCreate = 'User Create'
}

export enum ChartType {
    line = 'line',
    bar='bar'
}

export enum DirectionEnum {
  Top='top',
  Bottom= 'bottom'
}